Consolidated skill sets sql.
