+++ src/lib/db/resume/skillSets/sql.ts
@@ -4,6 +4,10 @@ import { db } from "@/lib/db"
 import { skillSets, skills } from "@/lib/db/schema"
 import type { SkillSet } from "@/types/resume"
 
+//
+// Skill sets.
+//
+
 // Find skill sets by candidate ID.
 export async function findSkillSetsByCandidateId(candidateId: string) {
 	// Select skill sets and skills.
@@ -63,4 +67,106 @@ export async function findSkillSetsByCandidateId(candidateId: string) {
 	}
 
 	return skillSetsObject
+}
+
+//
+// Skill set.
+//
+
+// Find skill set by candidate ID and skill set ID.
+export async function findSkillSetByCandidateIdAndSkillSetId(
+	candidateId: string,
+	skillSetId: string,
+) {
+	// Select skill set and skills.
+	const skillSetRows = await db
+		.select({
+			candidateId: skillSets.candidateId,
+			skillSetId: skillSets.skillSetId,
+			skillSetType: skillSets.skillSetType,
+			sortOrder: skillSets.sortOrder,
+			skillId: skills.skillId,
+			skill: skills.skill,
+			skillSortOrder: skills.sortOrder,
+		})
+		.from(skillSets)
+		.leftJoin(
+			skills,
+			and(
+				eq(skills.candidateId, skillSets.candidateId),
+				eq(skills.skillSetId, skillSets.skillSetId),
+			),
+		)
+		.where(
+			and(
+				eq(skillSets.candidateId, candidateId),
+				eq(skillSets.skillSetId, skillSetId),
+			),
+		)
+		.orderBy(skills.sortOrder)
+
+	// If there’s no skill set, return null.
+	if (skillSetRows.length === 0) return null
+
+	// Convert the database rows to an object.
+	const firstSkillSetRow = skillSetRows[0]
+	const skillSetObject = {
+		candidateId: firstSkillSetRow.candidateId,
+		skillSetId: firstSkillSetRow.skillSetId,
+		skillSetType: firstSkillSetRow.skillSetType,
+		sortOrder: firstSkillSetRow.sortOrder,
+		skills: skillSetRows
+			.filter(
+				(skillRow) =>
+					skillRow.skillId !== null &&
+					skillRow.skill !== null &&
+					skillRow.skillSortOrder !== null,
+			)
+			.map((skillRow) => ({
+				skillId: skillRow.skillId,
+				skill: skillRow.skill,
+				sortOrder: skillRow.skillSortOrder,
+			})),
+	}
+
+	return skillSetObject
+}
+
+//
+// Skills.
+//
+
+// Get skill by candidate ID, skill set ID, and skill ID.
+export async function findSkillByCandidateIdSkillSetIdAndSkillId(
+	candidateId: string,
+	skillSetId: string,
+	skillId: string,
+) {
+	// Select skill set and skill.
+	const [skill] = await db
+		.select({
+			candidateId: skillSets.candidateId,
+			skillSetId: skillSets.skillSetId,
+			skillId: skills.skillId,
+			skill: skills.skill,
+			sortOrder: skills.sortOrder,
+		})
+		.from(skillSets)
+		.innerJoin(
+			skills,
+			and(
+				eq(skills.candidateId, skillSets.candidateId),
+				eq(skills.skillSetId, skillSets.skillSetId),
+			),
+		)
+		.where(
+			and(
+				eq(skillSets.candidateId, candidateId),
+				eq(skillSets.skillSetId, skillSetId),
+				eq(skills.skillId, skillId),
+			),
+		)
+		.limit(1)
+
+	return skill ?? null
 }
